MANTA-3959 read from a random shark
To more evenly spread read load, shuffle the set of sharks to read from. During
the work I noticed there were two implementations of shuffle in the code, one in
picker.js, and one in utils.js. I couldn't find a call to the utils.js shuffle
in the code. The shuffle in picker.js has been proven in production. I moved it
to utils so that write and read share a common shuffle impl.

NOTE: TODO: there were 6 failing tests BEFORE and AFTER the change (same 6) that
I still need to investigate.